backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
